Added video popup
