+++ app/helper/model.ts
@@ -1,0 +1,5 @@
+export interface VideoData {
+    videoUrl: string,
+    isOpen: any, // eslint-disable-line @typescript-eslint/no-explicit-any
+    onClose: any // eslint-disable-line @typescript-eslint/no-explicit-any
+}
