+++ components/homeSections/podCasts/data/podcastData.json
@@ -3,27 +3,27 @@
     {
       "id": "1",
       "imageUrl": "/images/podcasts/senatorsprogram.jpg",
-      "link": "https://youtu.be/LNuesj6LeRc?si=UKIT24XkTjXrMuwT"
+      "link": "LNuesj6LeRc"
     },
     {
       "id": "2",
       "imageUrl": "/images/podcasts/marketconsulprogram.jpg",
-      "link": "https://youtu.be/KIlkhmxTi5M?si=poTYNVTme_EVMuiI"
+      "link": "KIlkhmxTi5M"
     },
     {
       "id": "3",
       "imageUrl": "/images/podcasts/networkconcordatprogram.jpg",
-      "link": "https://youtu.be/D0WFAAlEGOs?si=KgMMvTpWtlIJR_17"
+      "link": "D0WFAAlEGOs"
     },
     {
       "id": "4",
       "imageUrl": "/images/podcasts/diplomatprogram.jpg",
-      "link": "https://youtu.be/plMKga9xAqE?si=VgvKq-fav-jJe2oI"
+      "link": "plMKga9xAqE"
     },
     {
       "id": "5",
       "imageUrl": "/images/podcasts/ambassadorprogram.jpg",
-      "link": "https://youtu.be/gffjTUdgfDI?si=P4MY5r5N7c9zhMGT"
+      "link": "gffjTUdgfDI"
     }
   ]
 }
