+++ components/homeSections/podCasts/types.ts
@@ -4,4 +4,5 @@ export interface Podcast {
     description: string;
     imageUrl: string;
     link: string;
+    onOpenVideoPopup: any; // eslint-disable-line @typescript-eslint/no-explicit-any
 }
